+++ lib/get_arch_core.dart
@@ -6,6 +6,7 @@ export 'domain/i_failure.dart';
 export 'domain/i_entity.dart';
 export 'domain/env_config.dart';
 export 'domain/extension.dart';
+export 'domain/live_data.dart';
 export 'package:dartz/dartz.dart' show Either, Unit;
 export 'package:rxdart/rxdart.dart';